backend/requirements.txt
Flask==3.0.2
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
pymongo==4.6.2
bcrypt==4.1.2
requests==2.31.0
python-dotenv==1.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
numpy==1.26.4
Pillow==10.2.0
nibabel==5.2.0 